+++ relay-event-normalization/src/normalize/span/tag_extraction.rs
@@ -77,6 +77,7 @@ pub enum SpanTagKey {
     AppStartType,
     ReplayId,
     CacheHit,
+    CacheKey,
     TraceStatus,
     MessagingDestinationName,
     MessagingMessageId,
@@ -124,6 +125,7 @@ impl SpanTagKey {
             SpanTagKey::FileExtension => "file_extension",
             SpanTagKey::MainThread => "main_thread",
             SpanTagKey::CacheHit => "cache.hit",
+            SpanTagKey::CacheKey => "cache.key",
             SpanTagKey::OsName => "os.name",
             SpanTagKey::AppStartType => "app_start_type",
             SpanTagKey::ReplayId => "replay_id",
@@ -464,6 +466,11 @@ pub fn extract_tags(
             {
                 let tag_value = if *cache_hit { "true" } else { "false" };
                 span_tags.insert(SpanTagKey::CacheHit, tag_value.to_owned());
+            }
+            if let Some(cache_keys) = span.data.value().and_then(|data| data.cache_key.value()) {
+                if let Ok(cache_keys) = serde_json::to_string(cache_keys) {
+                    span_tags.insert(SpanTagKey::CacheKey, cache_keys);
+                }
             }
         }
 
@@ -1559,6 +1566,7 @@ LIMIT 1
                         "trace_id": "77aeb1c16bb544a4a39b8d42944947a3",
                         "data": {
                             "cache.hit": true,
+                            "cache.key": ["my_key"],
                             "cache.item_size": 8,
                             "thread.id": "6286962688",
                             "thread.name": "Thread-4 (process_request_thread)"
@@ -1570,13 +1578,14 @@ LIMIT 1
                         "timestamp": 1694732409.3145,
                         "start_timestamp": 1694732408.8367,
                         "exclusive_time": 477.800131,
-                        "description": "get my_key",
+                        "description": "mget my_key my_key_2",
                         "op": "cache.get_item",
                         "span_id": "97c0ef9770a02f9d",
                         "parent_span_id": "9756d8d7b2b364ff",
                         "trace_id": "77aeb1c16bb544a4a39b8d42944947a3",
                         "data": {
                             "cache.hit": false,
+                            "cache.key": ["my_key", "my_key_2"],
                             "cache.item_size": 8,
                             "thread.id": "6286962688",
                             "thread.name": "Thread-4 (process_request_thread)"
@@ -1595,6 +1604,7 @@ LIMIT 1
                         "trace_id": "77aeb1c16bb544a4a39b8d42944947a3",
                         "data": {
                             "cache.hit": false,
+                            "cache.key": ["my_key_2"],
                             "cache.item_size": 8,
                             "thread.id": "6286962688",
                             "thread.name": "Thread-4 (process_request_thread)"
@@ -1626,6 +1636,26 @@ LIMIT 1
         assert_eq!(tags_1.get("cache.hit").unwrap().as_str(), Some("true"));
         assert_eq!(tags_2.get("cache.hit").unwrap().as_str(), Some("false"));
         assert_eq!(tags_3.get("cache.hit").unwrap().as_str(), Some("false"));
+
+        let keys_1 = Value::Array(vec![Annotated::new(Value::String("my_key".to_string()))]);
+        let keys_2 = Value::Array(vec![
+            Annotated::new(Value::String("my_key".to_string())),
+            Annotated::new(Value::String("my_key_2".to_string())),
+        ]);
+        let keys_3 = Value::Array(vec![Annotated::new(Value::String("my_key_2".to_string()))]);
+        assert_eq!(
+            tags_1.get("cache.key").unwrap().as_str(),
+            serde_json::to_string(&keys_1).ok().as_deref()
+        );
+        assert_eq!(
+            tags_2.get("cache.key").unwrap().as_str(),
+            serde_json::to_string(&keys_2).ok().as_deref()
+        );
+        assert_eq!(
+            tags_3.get("cache.key").unwrap().as_str(),
+            serde_json::to_string(&keys_3).ok().as_deref()
+        );
+
         assert_debug_snapshot!(measurements_1, @r###"
         Measurements(
             {