+++ CHANGELOG.md
@@ -55,6 +55,8 @@
 - Support new cache span ops in metrics and tag extraction. ([#3598](https://github.com/getsentry/relay/pull/3598))
 - Extract additional user fields for spans. ([#3599](https://github.com/getsentry/relay/pull/3599))
 - Disable `db.redis` span metrics extraction. ([#3600](https://github.com/getsentry/relay/pull/3600))
+- Extract status for spans. ([#3606](https://github.com/getsentry/relay/pull/3606))
+- Extract cache key for spans. ([#3631](https://github.com/getsentry/relay/pull/3631))
 
 ## 24.4.2